move interfaces into src
