+++ src/pages/api/events/index.ts
@@ -1,7 +1,7 @@
 import { NextApiRequest, NextApiResponse } from "next"
 import { MongoClient, Db } from "mongodb"
 
-import { EventDocument } from "../../../../interfaces"
+import { EventDocument } from "../../../interfaces"
 
 let cachedDb: Db
